08/07
lista colectivos en sistema

+++ tp final linea colectivos (Correcciones de Franco)/Project34/cColectivo_nuevo.h
@@ -14,19 +14,15 @@ protected:
 	bool aire;
 	bool direccion_electrica;
 	int CantidadPersonas;
-	//cRamal *ramal;
-	//int CA;
-	//int CAPMAX; Ya lo heredo de la clase cColectivo y vale 60.
-	/*cPersona**Lista_Personas;*/
-
+	
 
 public:
 	cColectivo_nuevo();
 	cColectivo_nuevo(const string _codigo, bool _estado, cRamal * ramal, bool _aire, bool _direccion);
-	/*void Abrir_Puerta();*/
+	
 	void Asignar_Nuevo_Ramal(cRamal * ramal1, cRamal * ramal2, cRamal * ramal3);
 	void Bajar_Personas();
-	float Cobrar_Boleto();
+	double Cobrar_Boleto();
 	void SistemaGPS();
 	void Subir_Personas();
 	bool Verificar_Capacidad();
@@ -34,7 +30,6 @@ public:
 	//Devuelvo un contador que guarda la cantidad de personas que se subieron en el colectivo.
 	int getCantidad_de_Personas() { return CantidadPersonas; }
 
-	/*void set_CAPMAX(int capacidad);*/
 	int get_CAPMAX() { return CAPMAX; }
 
 	//No estoy seguro si se puede devolver asi el const en el .h asi.
@@ -43,6 +38,6 @@ public:
 	friend ostream& operator<<(ostream& o, cColectivo_nuevo& colec);
 	virtual ~cColectivo_nuevo();
 
-
 };
+
 ostream& operator<<(ostream& o, cColectivo_nuevo& colec);
